feat: Improve Contact and Projects pages
The Contact page now includes a cursor text component for the message link, improving the user experience. Additionally, the Projects page has been updated to include a flex-1 class, ensuring it takes up the available space on the screen.

This commit also includes a minor update to the Template component, adding a footer component to the page layout.

#